requirements.txt
pandas==1.3.5
mysql-connector-python==8.0.27
mysqlclient==2.1.0
slackclient
yfinance